requirements/base.txt
Django==1.9.1
django-braces==1.4.0
django-model-utils==2.0.3
django-pure-pagination==0.2.1
Pygments==1.6
logutils==0.3.3
Django-Select2==5.5.0
django-extensions==1.6.1
webtest
django-webtest==1.7.8
https://github.com/vandersonmota/model_mommy/archive/master.zip
django-crispy-forms==1.5.2
django-constance==1.1.1
django-picklefield
django-storages
django-allauth==0.24.1
django-avatar
django-autofixture
django-filter
git+https://github.com/fernandoPardo/django-enhanced-cbv
https://github.com/yrik/django-allauth-multiple/archive/master.zip
sendgrid-django
pyopenssl
ndg-httpsclient
pyasn1
fabric
django-markdown==0.8.4
https://github.com/WeFindX/django-money-rates/archive/django-1.9-fixes.zip
fred==3.1
geoip2==2.2.0
py-moneyed
